separated structs for css parser
