+++ components/css/src/parser/mod.rs
@@ -1,6 +1,9 @@
+mod structs;
+
 use std::env;
 use io::output_stream::OutputStream;
 use super::tokenizer::token::Token;
+use structs::*;
 
 fn is_trace() -> bool {
     match env::var("TRACE_CSS_PARSER") {
@@ -35,178 +38,6 @@ pub struct Parser {
     reconsume: bool,
     /// Current token to return if being reconsumed
     current_token: Option<Token>
-}
-
-#[derive(Debug, PartialEq)]
-pub enum Rule {
-    QualifiedRule(QualifiedRule),
-    AtRule(AtRule)
-}
-pub type ListOfRules = Vec<Rule>;
-
-pub enum DeclarationOrAtRule {
-    Declaration(Declaration),
-    AtRule(AtRule)
-}
-
-/// A temporary stylesheet to store rules before
-/// transforming it into CSSStyleSheet
-pub struct StyleSheet {
-    pub rules: ListOfRules
-}
-
-/// A simple block
-/// https://www.w3.org/TR/css-syntax-3/#simple-block
-#[derive(Clone, Debug, PartialEq)]
-pub struct SimpleBlock {
-    /// Associated token (either a <[-token>, <(-token>, or <{-token>)
-    token: Token,
-    /// Block value
-    value: Vec<ComponentValue>
-}
-
-/// Function
-/// https://www.w3.org/TR/css-syntax-3/#function
-#[derive(Clone, Debug, PartialEq)]
-pub struct Function {
-    name: String,
-    value: Vec<ComponentValue>
-}
-
-/// QualifiedRule
-/// https://www.w3.org/TR/css-syntax-3/#qualified-rule
-#[derive(Debug, PartialEq)]
-pub struct QualifiedRule {
-    prelude: Vec<ComponentValue>,
-    block: Option<SimpleBlock>
-}
-
-/// AtRule
-/// https://www.w3.org/TR/css-syntax-3/#at-rule
-#[derive(Debug, PartialEq)]
-pub struct AtRule {
-    name: String,
-    prelude: Vec<ComponentValue>,
-    block: Option<SimpleBlock>
-}
-
-/// Declaration
-/// https://www.w3.org/TR/css-syntax-3/#declaration
-pub struct Declaration {
-    name: String,
-    value: Vec<ComponentValue>,
-    important: bool
-}
-
-/// ComponentValue
-/// https://www.w3.org/TR/css-syntax-3/#component-value
-#[derive(Clone, Debug, PartialEq)]
-pub enum ComponentValue {
-    PerservedToken(Token),
-    Function(Function),
-    SimpleBlock(SimpleBlock)
-}
-
-impl QualifiedRule {
-    pub fn new() -> Self {
-        Self {
-            prelude: Vec::new(),
-            block: None
-        }
-    }
-
-    pub fn set_block(&mut self, block: SimpleBlock) {
-        self.block = Some(block);
-    }
-
-    pub fn append_prelude(&mut self, value: ComponentValue) {
-        self.prelude.push(value);
-    }
-}
-
-impl AtRule {
-    pub fn new(name: String) -> Self {
-        Self {
-            name,
-            prelude: Vec::new(),
-            block: None
-        }
-    }
-
-    pub fn set_block(&mut self, block: SimpleBlock) {
-        self.block = Some(block);
-    }
-
-    pub fn append_prelude(&mut self, value: ComponentValue) {
-        self.prelude.push(value);
-    }
-}
-
-impl SimpleBlock {
-    pub fn new(token: Token) -> Self {
-        Self {
-            token,
-            value: Vec::new()
-        }
-    }
-
-    pub fn append_value(&mut self, value: ComponentValue) {
-        self.value.push(value);
-    }
-}
-
-impl Declaration {
-    pub fn new(name: String) -> Self {
-        Self {
-            name,
-            value: Vec::new(),
-            important: false
-        }
-    }
-
-    pub fn append_value(&mut self, value: ComponentValue) {
-        self.value.push(value);
-    }
-
-    pub fn last_values(&self, len: usize) -> Vec<&ComponentValue> {
-        self.value.iter().rev().take(len).collect()
-    }
-
-    pub fn last_token(&self) -> Option<(usize, &Token)> {
-        for (index, value) in self.value.iter().rev().enumerate() {
-            if let ComponentValue::PerservedToken(token) = value {
-                return Some((index, token));
-            }
-        }
-        return None;
-    }
-
-    pub fn pop_last(&mut self, len: usize) {
-        for _ in 0..len {
-            self.value.pop();
-        }
-    }
-
-    pub fn remove(&mut self, index: usize) {
-        self.value.remove(index);
-    }
-
-    pub fn important(&mut self) {
-        self.important = true;
-    }
-}
-
-impl Function {
-    pub fn new(name: String) -> Self {
-        Self {
-            name,
-            value: Vec::new()
-        }
-    }
-
-    pub fn append_value(&mut self, value: ComponentValue) {
-        self.value.push(value);
-    }
 }
 
 impl Parser {
